backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
chromadb==0.4.22
numpy<2.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
aiofiles==23.2.1
